This version mostly works, but it has some aspects that do not function optimally.
It works well if all you do is type characters, backspace,
up/down-arrow, escape.  As soon as you start changing the highlight or
cursor position, things get a little weird at times.  Editing ostly
happens as one would expect (except when deleting from the end after a
manual selection of a string that goes to the end).  Autocomplete
updates will happen at unexpected times or not at all (when you would
expect it to happen).  However, these behaviors are not obtrusive and
are manageable, so I am checking this version in in its own branch.

However, I learned about the interface as I manipulated it - call it a
trial run to figure things out.  Now that I have an understanding of how
it works, I have a better idea of how to implement it in a way that will
work more fluidly.  Let me make some notes here:

The values associated with the text field have a before & after edit
state.  keyPressed reveals the state before edit resulting from a typed
key.  keyTyped allows you to figure out if there has been text changed
(except when using cut/paste).  keyReleased is where manipulations need
to happen - and the text field is updated automatically between keyTyped
and keyReleased, so this is where you can change it.  The data to track
for the state before and after are:

text in the text field
length of the text in the text field
selected text
selection start
selection end
selected index (in the popup menu)

Before this edit, all the backspace did was select text - it wouldn't
update the selected index and it was very difficult to manually edit the
text in the text field to enter wild card characters (?,*).  All I'd
like to do is, if the current non-autocomplete text matches something,
update the match and highlight the autocompleted text.  If it doesn't
match anything, don't autocomplete and let the user edit the text
(instead of highlight when backspace is typed).  The popup menu should
be shown and hidden based on whether there's a match.