Put in a temporary patch (with a TODO comment) to fix an issue with the color sliders not getting the value that was entered from a double-click.  Turns out that updateThumbPos expects a value between 25-475, but the value sent in from the entry box was calculated between 0-450.  The mouse drag of the sliders sends in adjusted values (25-475), so incrementing the value sent in in the function didn't work.  I put a static addition of 25 in the call to the function because I didn't know how to access the min in the context from where the function was being called.
I also fixed some logic that was causing another bug I found during
testing.  If the user tried to set the max value using the number entry
method, the slider would not change.  The case was being missed because
the check of the value used a '<' instead of a '<='.  To be careful, I
added a check on equality to the max because nextPos could actually be
another slider.