Added the selection to the overview panel and while I was there, I improved the drawing of circles around small selections.
The decision to draw a circle is now based on the size of the selection
in pixels as opposed to a single cell being selected.  Multiple circles
will now also get drawn if multiple small selections exist.  It also
prevents the drawing of overlapping circles if selections are close
together.  All selection borders are now yellow (as opposed to the
circle being drawn in white before).