Enhanced the selection of whole rows & columns.
Shift-click now toggles the selection from the col/col clicked to the
closest row/col of the opposite state (selected or unselected).

Command-click now toggles the selection of a single row/col.

Regular click deselects everything and selects the row/col clicked on.

There remains a bug however with the rows. The GlobalMatrixView
correctly updates row selections, but not the InteractiveMatrixView.