This version has fixed a bug where the JPanel inside the JScrollPane was not updating its size from the previous zoom step and thus, when the font size changed, either the longest label could not be fully read (after a zoom-in step) or there was extra space after the labels (after a zoom out step).  All that was needed was a call to setPreferredSize based on the longest string.
I enabled the long scrollbar for each of the label jpanels, but only on
hover, in preparation for the new functionality I have planned.
Currently on hover, it is set to always show for debug purposes, but I
intend to set to as needed once I have all labels populating the class.

I also fixed a few bugs where the default minimum label size was not
getting set in the label settings dialog.  I may have tweaked the layout
in DendroView.