Still a lot to do on the label port, so this is a rough draft just to get an idea of how it would look & feel.  Also made a couple fixes to the saving of font sizes since it was annoying to have to fix all the time.
Currently, the labels are not drawn correctly when zoomed in enough to
display labels as usual.  The label scrollbars do not work and hovering
over them causes the label pane content to go away.  The blue circle is
not drawn in the correct position.  Scrolling on the large data is
really slow.  Label scroll position is not updated on mouseEntered (but
updates correctly later on).  Scrolling the label port doesn't update
the labels drawn immediately.  Moving the divider is 1 rare case moves
the labels slightly.