Fixed a big that was causing the matrix hover indicator to be drawn in a slightly inaccurate position.
When the column label pane width was not evenly divisible by the font
size (+ SQUEEZE), the modulus had to be added to the offscreen width so
that the position of the indicator could be accurately shown.  It turned
out that I had the viewport height and width swapped in the calculation
of that modulus.