Added a gap between the label panes and the matrix.  Commented out a lot of label port code.  Created a scrolling label pane where the label of the cell currently hovered over is directly adjacent to the cursor and all other labels are pushed over based on the minimum font size.
This is actually my original idea about the labels zooming about
following the cursor.

I encountered what might be a bug: the getFirstVisible() method was
returning an incorrect data index.  I switched the logic to rely on
getIndex(pixel coordinate) instead, but that just side-stepped the
issue.  I wonder if this is an order of events issue...