Finished the first approved version of the label port implementation. All the bugs that are significant and fixable should be addressed.  The only annoying things I couldn't figure out was the precise width of the matrix hover indicator.  Plus, when you scroll the labels, the indicator bar briefly appears at the top of the scroll area.
A lot of things were implemented in this to get things to work well.
Here's a list of issues addressed in this commit. If it's marked with an
"X", it was nixed...

√ Draw indicator bar and hover indicator
√ hovering over top left cell makes all text bold
√ update hover index when dragging
√ Highlight labels during draw selection
√ Update the hover index during scrolling the scroll wheel
√ hovering over secondary scrollbar changes the primary hover position
to the beginning - should change to end because that's where the
scrollbar is - at the end of the label area
√ The label indicator breadth is drawn 1 pixel too long in the column
label pane.
X when bottom justified & manually scrolled to top, adjusting the
divider from the top to a midpoint shifts the scroll a couple pixels
√ change highlight colors:
	selection
		yellow IMV & GV border
		yellow text border/background
		yellow tree color
	label port
		blue IMV border
		blue text
		tree unaffected
		all goes away on hover? (e.g. labels disappear & border goes away)
		when both label ports are synched with the matrix, all blue goes back
to previous value
	hover
		red text
X		red tree highlight
X Hints are drawn in the wrong place when the matrix is zoomed in and
scrolled toward the beginning (left/top). Plus the secondary scroll bars
don't go away.  **** COULD NOT REPRODUCE
√ The knob/thumb/handle thing is the wrong size after dragging a
secondary scrollbar
√ Not a bug, but I need to implement the adjustmentValueChanged stuff
for the column labels (both in columnLabelView and in LabelView)
√ The secondary scroll position is not being rescrolled and the
scrollbar thumb is not being drawn the first time when hovering onto the
IMV. It corrects itself if the mouse stops for a second then moves again
or when the mouse is hovered over the label panes, or when the scroll
wheel is moved
√ scrolling over the labels does not update the hover highlight and the
scroll amount is not correct for the label area (probably updating the
hoverIndex will fix that anyway)?
√ scrolling over the matrix is not updating on the last scroll
increment. The matrix and labels do update if you click something or do
something that updates the image
√ Scrolling over IMV doesn't scroll the labels as much as it should
√ the red highlighted label is not updating when the matrix is scrolled
X Ability to drag indicator bar(s?)
√ Make scrolling scroll the matrix
√ secondary scrollbar doesn't work (and image goes away on hover over
scrollbar (because exited area))
√ mouseExited the IMV not turning off the hover animation
√ Clicking beside the thumb in the secondary scrollbar is not working
√ The secondary scroll position is not being remembered after the hint
is displayed
√ When justification is changed, reset the secondary scrollbar
X update data ticker on label hover (data ticker is going away/changing
to not hold labels)
X Blue circle not drawn in the correct position
X Scroll very slow on 6kx6k matrix
√ on hover over a label pane (from outside the window), the scroll
position is not remembered - though it corrects when you hover over the
IMV (or hover out & right back into the label pane) - I've determined
that everything's drawing correctly, so it seems to be an update issue
of some sort (like notify observers or something?
√ labels not refreshing on scrolls over the matrix (fixed similar label
pane scroll updates with paintImmediately)
X Have the label port follow the mouse
X Draw only the
X auto-scroll blue label port to (port larger than selection? center of
selected region : beginning of selected region) - minding the edges,
when a selection is made
√ Draw a partial string at the end, beyond the blue line when you can
see part of it
√ labels not drawn in the correct place when label port not drawn
√ label breadth indicator drawn 1 pixel too long
√ Top of some fonts are cut off, though the bar and indicator look
correct (look at diff between helv.Neue vs. courriernew)
√ when matrix scroll pos at end, the positioning of the strings is off
√ click on label is selecting wrong row/col
√ top and left justification is cutting off the justified side
√ left justified rows are not set up correctly when justification is
changed
√ left justified rows are not set up correctly on open
√ zoom hint uncentered and preferred size incorrect. They are undated
correctly later when you try to drag the scrollbar
√ bottom-justified col labels not updated properly after font size
change
√ bottom-justified col labels not set up properly on open
√ column secondary scroll doesn't work when top justified
√ highlight when hovering over labels is incorrect (it's highlighting
based on corresponding square in the matrix using the start of the label
pane (not the relative matrix position, i.e. it's not highlighting the
label for the row the mouse is next to, rather highlighting the label
for the row the mouse is hovering over in the label pane, considering
the offscreen portion of the label pane))
√ add a highlight of a label that corresponds to the hovered square in
the matrix
√ top of the top label is getting clipped off just slightly
√ label min size setting not remembered
X fixed size setting changing with min/max - looks like Chris programmed
it this way - as there's only 1 size saved in the prefs (as opposed to
current size and fixed size)
X scrollbar flashes as you hover over it
X scrolling (main) in the label pane shifts the labels back under the
main scrollbar (I think the main scrollbar is flashing: exit/enter
without shift)
√ labels not refreshing on scrolls in the label pane
√ zooming with the main zoom button is not going into the center
√ resizing the window is not updating the font size
√ add blue border
√ If you don't scroll all the way with the scroll wheel, the image is
not fully updated until you move the mouse - look at how GV updates
√ When data first opened, scroll to justification position
√ click on matrix is highlighting wrong labels
√ adjusting divider does not update the indicator