Implemented a feature where labels for every col/row are always at least partially visible.
Scrolling with the scroll wheel is a little jittery.  That wasn't
happening in the previous commit without this feature.  In fact, I had
fixed the jittery scrolling before by rearranging some code, but now the
jitteriness is back.  *sigh*

Also, scrolling by clicking on the scroll handle and dragging does not
update the pane correctly while dragging.