Fixed some issue with overrun arrows in the label panes.
There were some positions of the divider where it made the label pane
*just* smaller than the longest label and those arrows were not getting
drawn.  Fixed.

The overrun arrows were getting drawn in the wrong position after a
split pane divider adjustment.  This was due to the calculations using
scrollpane methods that had stale data and viewport dimensions had not
been updated.  Fixed.

Sometimes, when the viewport size exactly equaled the longest label, an
overrun arrow was needlessly drawn and scrolling would not move the
label, and put the overrun arrow on the other side.  Now that I think
about it, drawOverrunArrows should not be called, but I haven't done
that yet.  Fixed otherwise.

There are issues remaining that deceptively look like overrun arrow
issues, but they are not (e.g. the label start position being negative
and the overrun arrow always being drawn).