Finished the second draft of the "whizzing labels" implementation. Everything's a bit rough around the edges. Some encapsulation/compartmentalization is needed, but everything works, barring some unrelated bugs having to do with multiple data points under the same pixel.  Here's a list of the issues addressed:
Some labels not drawn when split pane divider is dragged to a position
between the longest & shortest label and in some cases, all labels are
completely out of view.

Resizing split pane larger (after very small) draws all labels and the
indicator bar as if the longest string was forced to the left edge even
though right-justified.

Overrun arrows in the row labels have backgrounds that cover 1 pixel of
text right of the left arrow, but 3 left of the right arrow.  It should
be 2 and 2 like the column label overrun arrows.

Overrun arrow backgrounds are not covering up underscores (or the
underscores are being drawn over them from a proximal field).

Overrun arrows not drawn at first when split pane is real small, until
secondary scroll occurs.

Column zoom/hover hint not centered when labels are scrolled to the top
or left side.

Zooming out from a non-label-port mode results in maxStrLen not coming
back down plus all sorts of side-effects. I think that the problem
relates to not enough updates upon font size change.

Indicator bar drawn in wrong position. Seems to happen in specific
divider bar positions, and only when the labels have been scrolled down
a little and the divider has been dragged up to the weird position.
Corrects by hovering over matrix or scrolling.

In some positions of the split pane divider when labels are
right-justified (only), you cannot scroll to the left all the way. I
suspect it has to do with the negative indent/offset from the indicator
bar.

Sometimes position indicator is not drawn when data is large.

Changing label type does not re-justify labels when viewing a cdt file.

Draw overrun arrows in non-label-port mode.

Adjusting the split pane divider shifts the labels when not in
label-port-mode.

Top & left justified labels in non-label-port-mode do not all scroll to
the correct 0 position - they run off the edge and you cannot scroll
them into view.  Probably that negative scroll value issue.

Indicator bar is in the middle of the pane and the labels run off the
top with no scrolling possible. Turned out that the font size wasn't
changed until after the max string size calculations were performed.

After dragging split pane divider up, short labels were not visible
(about a third) when top justified & fully zoomed out on small133x133
data. Hovering over the matrix fixes it. Looks like all the labels are
positioned as if the scroll bar was all the way at the bottom. This
turned out to be an improper update of the pane size in the secondary
dimension (length of max label string).