Eliminated a ton of code duplication by moving methods from the Column/RowLabelView classes into LabelView and simplifying the row/column label tracking in MapContainer.
All the contained lines of code inside methods moved to LabelView which
were different depending on whether it was a row/column label pane was
implemented as an abstract method in LabelView and a concrete method in
each subclass.

Changes to MapContainer served to eliminate row and column references
(since a particular container only knows about or is responsible for 1
type of label pane). All row and label variables and methods were boiled
down to generic label variables/methods without a row/column assignment.