Adjusted insets and gaps for components
The insets for the JPanel containing all main elements for DendroView have been adjusted to 3px (distance from window edge). The insets for GlobalOverviewPanel have been adjusted to 5px for the right and bottom sides of the panel to give it some distance from the main matrix and the label panels. Currently still trying to adjust gaps between navigation panels (scale buttons + matrix scrollbars) and the matrix which should be 0.

Resolves: -
See also: -