Documentation changes to separate developer from user documentation and to bring the documentation under the auspices of the change control provided by the repository.
The added documentation was first copied from the BitBucket project home
page and then edited.  The purpose of these edits was to start a
discussion on some guiding project principles on which we've had some
differences of opinion.  Discussing these formally in a document review
of the pull request these changes will generate is the actual intent of
these changes.  Hpefully, we will come to an agreed upon consensus.

Also, I upgraded the license from GPL 2 to 3 and removed outdated
documentation from TreeView 2.