Resolves issue #233.
I added a layer in front of seekAll.  Basically, all seekAll does now is
initiates the search in both the finder box it belongs to and the
companion's finder box (rows & columns together).  All the original code
from seekAll is now in seekAllHelper, which seekAll calls.

I had to add a handle into the companion finder box so that each finder
box can initiate the search in its partner's box.  I utilized existing
code.  Although the overall design to do it this way is not ideal, it
was easy to do and straightforward.  One MUST set the companion after
having instatiated 2 finder boxes by calling setCompanionBox.  That is
done in DendroView.

I added a couple new methods for convenience and readability:
getSearchTerm & isSearchTermEntered.