This resolves issue 234.  Searches are now initiated upon click of a combobox dropdown menu item.
Note, the method used in this update is non-standard.  Multiple means
and methods were tried.  The most reliable solution turned out to be not
straightforward.  If a better means is found to initiate a search from a
clicked menu item, this added code has been marked with a note about
removal.  The marked methods, classes, and specific lines (3 places
total) may be deleted.  For more details on the efforts surrounding this
issue, see this stackexchange thread:

http://stackoverflow.com/questions/33268726/how-do-you-detect-when-a-user-commits-a-jcombobox-selection-via-mouse-click-in-j