Resolves issue #211 - search box should retain focus after search.
I tried using isFocusOwner() on the searchTermBox's text field, but it
always returned false for some reason.  I'm not sure why that was, so I
simply set a boolean at the beginning and end of the search process (in
seekAll()) that DendroController can check before requesting that one of
the zoom buttons gets focus.  This allows us to to change focus to the
zoom button when either of the search term boxes has focus.

If anyone figures out how to properly check the search term box's focus
status, we should improve this solution.