Fixed a bug that omitted the last loop iteration when constructing the dropdown contents.  Also added logic to skip label types like EWEIGHT.
The logic for skipping EWEIGHT/GWEIGHT label types relies on the fact
that all of the values are the same.  Any label type whose values are
all identical are not added to the combobox simply because they do not
add any search power.  The exception is the selected visible label type.
If the user selects EWEIGHT and/or GWEIGHT as the visible label type,
those values will be forced to be present in the search box to avoid
confusion.