Replaced inline Strings with descriptive final String variables
In order to have a common point to edit Strings and actually understand their meaning within code, inline Strings have been declared as private final variables in CDTGenerator. This allows people to understand why a specific String is used in the code because the variable name should clarify what it actually stands for.