Added missing exception logs, TODO statements, made axisPrefix variable in HierCluster final
Some exception throughout the cluster code weren't logged via LogBuffer. While we want to implement a proper logging framework very soon, it's still preferable over a simple printStackTrace statement because more info is logged. Some parts of the code are very obviously obsolete or ripe for improvement, so I added some TODO statements on the fly to remember these code bits. The axisPrefix instance variable used to generate the tree file for whatever axis is being clustered should be final - it cannot be changed once an axis has been chosen for clustering in this class.