Made axis IDs in ClusterFileGenerator static for use elsewhere
The axis prefixes for the axis labels (ROW/ COL) are used in several classes so it is nice to let them be statically initialized in ClusterFileGenerator so that they can be used throughout the clustering code. These Strings (and similar ones in the same file) may be moved to a more general class in the near future.