Final adaptations for changing axis IDs from old GENE/ ARRY to new ROW/ COL
Changed the access modifier of the static variables describing axis ID names in ClusterController so that they can actually be accessed by other classes.
Changed the axis ID name references in HierCluster to set the correct labels to be used in ClusterFileGenerator and when generating the tree file (atr/ gtr) in HierCluster.
Adapted regex for recognizing common labels in PreviewLoader to look for the new axis prefixes.

Resolves: #214
See also: #321