Fixed problems resulting from the merge of the implementations of issues 180 & 211 with a master that was updated with conflicting code from other merges.
Search term color indication of red to denote "not found" was switched
from a red foreground color to a red background color which is retained
even upon highlight and is updated immediately (instead of upon focus
change).

Focus of the search box was retained by commenting out a request for
focus upon mouseEntered in IMV.  However, this breaks the intent of that
focus change: to allow keyboard navigation enabling upon mouseEntered.
It is recommended to address that issue via another means that is more
explicit: either an actual click or tab keypress or some other fancier
means of detecting the user's intent to navigate with the keyboard.