First rough prototype for asking a user if the row axis should be clustered
On the way to implement the expected behavior for files with previously existing clustered axes, the first prototype that checks the row axis for hints of previous clustering and asks the user how to proceed, has been implemented. Still very rough around the edges and the analog implementation for column clustering is missing. The crucial part of storing a new tree file for the newly clustered file has not been started yet.