Added logical check for tree file presence
When an axis is considered to be clustered, whether by clustering in the current session or by previous clustering with the tree file already loaded, it should be recognized as such. Before a newly clustered file is loaded, there is now a check for the presence of tree files for all axes that are considered to be clustered. If an axis, that is considered to be clustered, does not have a new tree file (new file name root + tree file extension), then the code recognizes this and also checks the old file set (same directory, the currently loaded pre-cluster file) for available tree files. In a future code change any such present tree file will be copied and added to the file set of the newly clustered file. This way it will be associated with the new file and loaded properly.