Repaired the highlight on hover functionality.
Had to create a repaintHoveredNode method in TRView which only paints
the hovered node to call from updateBuffer in ColumnTreeView.  The only
other method which did this was setHoveredNode which is called from
mouseMoved.  However, the purpose of that function was to paint over a
previously hovered node and then paint the given new hovered node.
Since the animation of the tree moving initiates repaints independent of
mouseMoved (because the mouse can be hovering over different components)
and because we don't need to determine the hoveredNode, we need to
update the image in the updateBuffer calls.