Fixed the alignment issues of the tree aligning to the labels.  this also addresses the choppiness issue as a bonus.
This solution takes advantage of the way the distance between tree
leaves is already calculated by supplying the width of the pane minus
the width of the incomplete labels drawn on each edge.  MapContainer
needed to add 2 new datamembers with setters & getters to keep track of
the width of the partial labels drawn, i.e. the offset of the area that
will be used to calculate tree leaf positions.

Note, this version has some debug prints that are turned on still.