Added the ability to hover over a data-linked tree, (i.e. hover over the tree without it being linked to the whizzing labels).
If cursor enters the tree pane from a data-linked / non-label-linked
state, the tree is kept visually linked to the data/static.  The the
cursor enters the tree pane while it is in a label-linked state (i.e.
non-data-linked state), the tree remains dynamically linked to the
whizzing labels, and thus moves with the cursor.

This is achieved by keeping track of the way the tree was last drawn and
checking that state upon mouseEntered in the tree pane.  Two data
members in MapContainer were added along with their associated
setters/getters:

lastTreeModeGlobal
keepTreeGlobal

I used the term "global" instead of "DataLinked" because I thought it
would be more intuitively understood.