Added the ability for the app to remember whether the user last interacted with the tree in a data-linked mode or with it in a label-linked mode.
Hovering over the tree when it is in data-linked mode will cause the
tree to stick in the data-linked mode until the user hovers over the
IMV.

Likewise, hovering over the IMV will cause the tree to stick in
label-linked mode until the user hovers off a label-linked pane (causing
the tree to temporarily switch to a data-linked view) AND hovers onto
the tree first.

Rule of thumb: hovering the mouse from the IMV up will keep things in
label-linked mode until it goes above the app.  Hovering the mouse down
from above will keep things in data-linked mode until it hits the IMV.
The intent is for the user to figure out intuitively how to switch
between the views.