Changed the highlighted subtree on hover to default to the parent node of the currently hovered branch.
If you're hovering over a leaf, the subtree of its immediate parent is
highlighted.  As you drag the cursor higher directly over that leaf,
once you pass the immediate parent, the parent higher up then becomes
the root of the highlighted subtree.

The previous behavior was only to highlight the leaf and select parents
only once you'd gotten above them.  This behavior was 1. unnecessary
since we already have a means of selecting single columns/rows using the
labels and 2. was somewhat difficult when selecting subtrees close to
the root because the tree switches from whizzing mode to global mode
once you hover above the root.