Linked the blue branch color as being associated with which labels are currently visible.  There are some caveats with this implementation.
There are as many as 3 pieces of information to convey: selected,
hovered (node & leaf), and visibility of the label.  Blue line = label
is visible.  Red line = hovered node (bold red line = hovered leaf).
Yellow outline = selected.

The problem is that hovered-node trumps visible label.  Can't see both
of those pieces of information at the same time.  If I made the red an
outline, then it would conflict with the yellow selected outline.  So I
really need a third way to denote the data.

Lance suggested that the blue be associated with a bar the same way it
is with the labels.  Just put a bar under the tree.  However, I'm not
keen on reproducing that code since it's rather complicated and linked
to label length.  I think it might be easier to just color the
background.  I do like the idea of drawing the blue as being
unassociated/disconnected with the tree because it's really just a range
anyway.  It would be better to draw the selections that way.

The next commit will undo a lot of what was done in this commit to use
the background color for what's visible label-wise.