Added a selectAddedIndexRange method so that shift-clicks and shift-click-drags on labels would only add to an existing selection instead of deselect everything first.
In a previous version of master, I had removed the call to
deselectAllIndexes from the selectIndexRange method because I thought
that that should be called explicitly, but it appears to have been
re-added, so I decided to make a separate function so that legacy calls
which depend on the deselection would still function as expected (though
I had previously checked for such legacy dependencies - however I just
assumed this time that new code must depend on it and just separated the
functionality in a new method).

I also altered the description of selectIndexRange to indicate that it
also deselects existing selections.  I think that in the future, we
should rename selectIndexRange so that it more accurately reflects what
it does, such as selectNewIndexRange, selectIndexRangeAnew,
selectFreshIndexRange - something that suggests that it does a
deselection first.