Changed the draw method for the tree to draw in a depth-first search pattern.  Also managed the drawing of the top hovered/selected node's dot so that it is the last thing drawn and will not get drawn over.
This was done by drawing the bottom of the tree before the top so that
selected internal nodes will be drawn better and so that when a node is
selected, the internal node will be visible.  The hovered/selected node
is passed back to the root so that after the root is drawn, the dot can
be drawn afterwards.  This is handled by obfuscating the draw method to
have a wrapper called draw and a helper called drawDFS (draw depth first
search).