Worked out logic flow for tree files and confirmation dialogs
This commit ensures that the dialog only pops up when expected and the file will be clustered as the user expects it. The tree file of an axis that has been clustered before will be transferred to the "new" file unless the user chooses to cluster it again. In that case a new file will be generated.