Added another case to check in tree file integrity test
If for some reason the current axis is considered to have been clustered before and the user does not want to explicitly recluster BUT no old tree file can be recovered for copying, then the matrix and labels will be kept in their clustered order. A dialog will pop up warning of the failed tree file recovery attempt alongside a message that no trees can be displayed as a result (no node/ link information available).

Resolves: #256

See also: -