Polished the generation of the PDF using freehep.
Changed the fillRect method with the drawRect method, which leaves no
gaps between the tiles (though PDF readers tend to draw this slightly
sloppily with some overlap that change the colors due to an alpha
channel).

Corrected the dimensions to correspond to the entire matrix.  This will
have to be changed to the current view at some point.

There is still a debugging PDF export button named "X" next to the zoom
buttons.  We will need to create a better integrated interface for this.