Polished some of the code for exporting images.
Added support for .ps and .svg, though the code that creates these files
is commented out.  I left it in there with a TODO comment.

There is currently no interface other than a quick button.  I added TODO
comments in DendroView to indicate where the testing button code was
added.