Renamed overallClusterStatus and clarified intent
The purpose of the overallClusterStatus member was unclear. Its intent is to track if an axis is onsidered to be clustered. This is important to validate tree file presence when clustering so that old tree files can be transferred, for example. The new naming more closely relates to the action in which the member is involved in.