Bug fix for replacing old axis IDs (gene / row) with newly ordered IDs
If a file is loaded that is already clustered, then previously old GIDs were taken over from the HeaderInfo of the old file. Row and col trees use specified identifiers (ROW..X, COL..X) as labels for their nodes. These have to match up with the IDs in the main cluster file or trees cannot be drawn properly. This commit ensures that labels will always use the same IDs as the tree files and ditch any other previous ID label.