Fixed bug where axis was reordered when it was only clustered in a previous session
When an axis was previously clustered, and the user opted out of clustering again, a flag for its status as clustered axis was set to true, which caused the clustering program to reorder this axis. The reordering was wrong because if the axis was already clustered, the order should not be changed. This did not affect the data matrix, but the axis ID label names. The old tree file would still use nodes with the old label IDs and end up completely scrambled with the newly ordered IDs.