Further implementation of ClusteredAxisData in ClusterFileGenerator
The ClusteredAxisData objects are now used more frequently in the ClusterFileGenerator class so that some code repetition can be avoided. The string representation cdt_s of the data matrix has been removed and replaced with a more local variant. Now only a single row is transformed into Strings and sent to the buffered writer for saving.