Fixed String size issue in BoundaryBox and renamed PreferencesMenu to LabelSettings
The String of the left BoundaryBox for the ColorPicker was drawn smaller
than the right one. This could be fixed by explicitly setting the font
before drawing the string.

PreferencesMenu has been renamed into LabelSettings, which is far more
fitting for its current purpose (it evolved from a multi-options window).

Fixed: #343
See also: -