Improved cancellation process
When the user clicked cancel, the UI behaved as it should but there were
some bad concurrency issues. Especially deleting files and folders after
unsuccessful clustering would not work properly. The issues are not
completely fixed yet, as a CancellationException is still occurring and
must be handled after some research about the topic. The app is behaving
properly otherwise.