Added export dialog and adjusted CustomDialog
Some classes depending on CustomDialog were adjusted. The main task was to
implement a simple export dialog with a few options and a preview panel.
In order to realize the preview panel, a BufferedImage from every relevant
panel (trees, matrix) has been created. The matrix already provided a
proper BufferedImage while a getSnapshot() message needed to be
implemented for the trees. The BufferedImages are then drawn to JPanels,
which are passed to the new dialog.

The tree getSnapshot() method currently does NOT take into account the
preview panel dimensions when creating the image because updateBuffer() is
implictly used through paint(). There needs to be a special method that
draws the visible trees in the provided dimensions.