Implemented the ability to set the aspect ratio to that of what's on the screen and added the ability to set the tree height dynamically based on the size of the data.  Added the ability to export either the entire matrix, what's visible, or what's selected.
I have not yet incorporated the ability to show the selection in the
output, not have I added the ability to show selections in the output
image.

The tree size is based on the longest final dimension of the image
(default 20%).

Trees are drawn up to the point where they run out of the designated
tree area now.

The code is cleaned up fairly well, although some methods still need
headers.