Implem. correctly scaled trees for export preview
The trees will be scaled correctly to the panels in the export preview.
What the user sees on the screen is what will be in the preview and what
will be exported (for now). To achieve this, the getSnapshot() method was
moved from TRView to its children, RowTreeView and ColumnTreeView in order
to achieve specific scaling with the primary and secondary LinearTransform
objects and the destination rectangle.