Implemented PR changes and fixed panel sizes
The changes suggested in PR #39 have been applied (except more paper
types, list not available offline). A fix to the JPanel sizes in the
dialog has been applied as well. Previously the total size of the
previewPanel would be less than the combined size of its components. These
sizes are explicity determined in the MigLayout parameters.