Added the ability to "show selections" in the exported image.
Both trees and matrix rectangle outlines are drawn, but I opted not to
include circles, as I think that it is both unlikely to happen and
unlikely to be desired.

I implemented a new method is ArrayDrawer that is named the same as a
method in InteractiveMatrixView (findRectBoundaries), but it returns
data indexes as boundaries instead of pixel indexes.  The code was also
streamlines/simplified slightly.

There is a very minor issue with the exported yellow tree outlines  the
corners have a little gap in vector output formats.  This is because I
used drawPolyline instead of drawLine.  drawLine was used for the black
lines of the trees because there was a weird erroneous behavior where
sometimes the middle point was skipped.  For whatever reason, this does
not seem to happen with the yellow lines.  It might be worth a try at
some point in the future to return to drawing polylines for the tree
branches to smooth out those corners.

I tried to simplify some of the code in TRView, but more could be done
to reduce redundancy.  I will create an issue to try and somehow merge
and simplify the drawing methods for the screen versus export, however,
some of the redundancy, I think, makes the code easier to follow and
understand, which is why I originally separated the drawing from
drawSingle and created drawLeftBranch and drawRightBranch.  I renamed a
bunch of variables to make that code easier to understand.